+++ app/src/main/res/layout/activity_main.xml
@@ -54,4 +54,13 @@
         android:layout_marginLeft="16dp"
         android:layout_marginRight="16dp"/>
 
+    <Button
+        android:id="@+id/goToSchoolPage"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="이동"
+        android:layout_below="@id/buttonSubmit"
+        android:layout_marginTop="16dp"
+        android:layout_marginLeft="16dp"
+        android:layout_marginRight="16dp"/>
 </RelativeLayout>
