+++ build.gradle
@@ -2,9 +2,11 @@
 plugins {
     id 'com.android.application' version '8.1.1' apply false
     id 'com.google.gms.google-services' version '4.4.0' apply false
+
 }
 
-//dependencies {
+dependencies {
+
     // Firebase Realtime Database를 사용하기 위한 종속성
     //implementation 'com.google.firebase:firebase-database:20.0.1'
 
@@ -15,4 +17,4 @@ plugins {
 
     // 필요한 경우 Firebase UI 종속성도 추가할 수 있습니다.
     // implementation 'com.firebaseui:firebase-ui-database:8.0.0'
-//}
+}
